refactor(gui): Rename and reorganize methods for clarity
- Renamed "refresh_grid" to "refresh_left_widget", as this
method refreshes the left widget, which can be either a grid
or tabs.
- Renamed "remove_grid" to "remove_left_widget" for similar
reasons.
- Removed "contained_elements" from possible values of
"current_view". "current_view" now only stores the name of
the current element being shown (e.g., "websites"), reflecting
previous refactoring.
- Removed "get_app_insertion_result" method; its sole purpose
was to return a value after adding an application, which was
simplified by replacing it with a lambda function.
- Removed an "elif" statement in "validate_input", as this method
is solely used for validating input when creating either a website
or environment.