fix(gui): Remove redundant grid attribute and refactor grid-update logic
- Fixed a bug that occurred when there were many contained elements and environments;
clicking an element sometimes failed to highlight it. The issue was due to using the
length of the grid attribute in main_view (which corresponded to the last created grid)
for loop iterations and label retrieval. By removing this attribute, the logic was
refactored in the controller to track the currently displayed grid for all operations.
- Simplified the style_clicked_label method using get_clicked_element_pos to check whether
the clicked label is the same as the previously clicked label (if one exists).
- Added docstrings to the style_clicked_method for better documentation.
- Created enable_buttons and disable_buttons methods to encapsulate the logic for enabling
or disabling the start and delete buttons across various parts of the program.