refactor: Add interface modules between CLI/GUI and database
- Created "app_validation" and "validations" modules in the
database package to serve as an interface between CLI
commands/GUI events and the database. This allows both of them
to perform the same operations on the database but handle the
results differently (e.g., showing an error window for GUI,
throwing a parser error for CLI).
- Refactored all commands, "main_controller", and "view_utils"
to use "app_validation" or "validations" instead of directly
interacting with the database or duplicating logic across GUI and CLI.