refactor(components): Create ui and dialogs packages
Previously, all UI components were stored in the ui module, and the same
applied to dialogs within the dialogs module. Although this approach worked,
it made changes or refactoring challenging as each file had approximately 500
lines of code.

To improve readability, maintainability, and ease future development or
refactoring,each component class in the ui and dialogs modules was moved to its
own file. This ensures that every component is now contained in its own separate
file.