refactor(View): Remove current_env argument from change_elem_view
The current_env parameter was removed from the change_elem_view method
because its value is obtained by calling GuiStateMachine. Passing it as
a parameter was redundant, as the method relies on GuiStateMachine to
return the correct value regardless of whether it's accessed directly or
passed through a parameter. This change simplifies the method's interface
and reduces unnecessary dependencies.